[BUGFIX] ContentCollection is wrapped in live workspace when logged in
Since we just check if the session has access to the backend we
incorrectly also wrap content collections being rendered in the live
workspace. This means that when visiting the frontend when logged in it's
being rendered differently than when not logged in. Since we only
need the wrapping for the backend interface we should skip wrapping
if in live workspace.

Change-Id: I6b8dfb9453a47aa0929f076525df8b23c361c837
Reviewed-on: https://review.typo3.org/25565
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann
Reviewed-by: Bastian Waidelich
Tested-by: Bastian Waidelich
Reviewed-by: Dominique Feyer
Tested-by: Dominique Feyer
Reviewed-by: Rens Admiraal
Tested-by: Rens Admiraal