[BUGFIX] ContentEditable is wrapped in live workspace when logged in
Since we just check if the session has access to the backend we
incorrectly also wrap content editables being rendered in the live
workspace. This means that when visiting the frontend when logged in it's
being rendered differently than when not logged in. Since we only
need the wrapping for the backend interface we should skip wrapping
if in live workspace.

Change-Id: I2f4b3a961b4a007031e590985ea9f9cffac7266e
Reviewed-on: https://review.typo3.org/25566
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann
Reviewed-by: Bastian Waidelich
Tested-by: Bastian Waidelich
Reviewed-by: Dominique Feyer
Tested-by: Dominique Feyer
Reviewed-by: Rens Admiraal
Tested-by: Rens Admiraal