[!!!][BUGFIX] Menu TypoScript object wrapped as content element
The menu TypoScript template is currently using the
contentElement view helper for wrapping the menu.
This is incorrect since the menu is not a node type,
but instead a plain TypoScript object.

This is a breaking change because it removes a wrapping div
from menus if you used the default template.
If your markup relies on the wrapping div, you can add it to
your page template(s) or use a custom menu TS object template.

Change-Id: I59d9f6a08c3583c70e67caeeefd406d39d36d846
Reviewed-on: https://review.typo3.org/25567
Reviewed-by: Bastian Waidelich
Tested-by: Bastian Waidelich
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann
Reviewed-by: Dominique Feyer
Tested-by: Dominique Feyer
Reviewed-by: Rens Admiraal
Tested-by: Rens Admiraal