[!!!][FEATURE] Make URL suffix configurable
Currently all node URLs created by Neos have a suffix ".html". Those
can only be changed by replacing the neos routes with custom routes.

With this change it's possible to adjust/disable the suffix only by
configuring the ``defaultUrlSuffix`` variable of the main route.

This is a breaking change because the new variable is *required*. If
you don't use the standard Neos distribution you thus need to adjust
the Neos route in your global ``Configuration/Routes.yaml``::

 -
   name: 'TYPO3Neos'
   uriPattern: '<TYPO3NeosSubroutes>'
   subRoutes:
     TYPO3NeosSubroutes:
       package: 'TYPO3.Neos'
       variables:
         'defaultUrlSuffix': '.html'

If you want to disable the suffix, replace ".html" by an empty string.

Change-Id: I2f2281a06c66dddf496b50a0cf2f3d05943c7041