[!!!][FEATURE] Make URI suffix configurable
Currently all node URLs created by Neos have a suffix ".html". Those
can only be changed by replacing the neos routes with custom routes.

With this change it's possible to adjust/disable the suffix only by
configuring the ``defaultUriSuffix`` variable of the main route.

This is a breaking change because the new variable is *required*. If
you don't use the standard Neos distribution you thus need to adjust
the Neos route in your global ``Configuration/Routes.yaml``::

 -
   name: 'TYPO3Neos'
   uriPattern: '<TYPO3NeosSubroutes>'
   subRoutes:
     TYPO3NeosSubroutes:
       package: 'TYPO3.Neos'
       variables:
         'defaultUriSuffix': '.html'

If you want to disable the suffix, replace ".html" by an empty string.

Change-Id: I14c8a98f341fbc312c57fee44cd383271778e57b
Reviewed-on: https://review.typo3.org/25746
Reviewed-by: Sebastian Kurfuerst
Tested-by: Sebastian Kurfuerst
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann
Reviewed-by: Bastian Waidelich
Tested-by: Bastian Waidelich