[TASK] Move lastVisitedNode functionality to external script
Replaces the inline script that stores the last visited document node
in the session storage to an external file so that it doesn't violate
the Content Security Policy specification and so that it can be cached
by the browser.

The current node identifier is passed to the external script via a
data-neos-node attribute on the script tag.

Change-Id: I7237f97499be1e56c3ea9fa0c4403c333149c220
Depends: Id3d887094e3a79e58ccfe4a89dff87ca0a96c291
Related: #40304
Reviewed-on: https://review.typo3.org/25688
Reviewed-by: Bastian Waidelich
Tested-by: Bastian Waidelich