[FEATURE] Improve contentElement.editable ViewHelper
This improves the EditableViewHelper so that it is easier to use and
less error-prone.

It mainly makes specification of the property value optional.
Before::

 <neos:contentElement.editable property="propertyName">
   {propertyName -> f:format.raw()}
 </neos:contentElement>

Now::

 {neos:contentElement.editable(property: 'propertyName')}

Note: This change is fully backwards compatible because any child nodes
of this ViewHelper overrule the specified property argument.
It is however recommended not to set them as that can lead to unexpected
results. It is also recommended to use Fluid's inline syntax for better
readability and to make browsers render the placeholder when opening the
template file directly.

Change-Id: I2e6e9035a85b58cd9f28f5f2f5b57b7631f64497
Related: #53046
Reviewed-on: https://review.typo3.org/25765
Reviewed-by: Bastian Waidelich
Tested-by: Bastian Waidelich
Reviewed-by: Christopher Hlubek
Tested-by: Christopher Hlubek