[BUGFIX] CreateJS: Handle modified Aloha editables correctly
This change updates CreateJS with a bugfix for the alohaWidget. The
widget did not correctly receive all changes to an Aloha editable,
because Aloha triggers no events for some changes like formatting
actions. These have to be polled explicitly.

The updated source in "update-createjs-to-master" can be changed back
to the main create repository as soon as our pull request is merged.

Change-Id: I7aee04f6657318c80202766aa9ac1771e7ed9018
Reviewed-on: https://review.typo3.org/26039
Reviewed-by: Sebastian Kurfuerst
Reviewed-by: Rens Admiraal
Tested-by: Rens Admiraal