[TASK] Re-introduce a view helper for manually wrapping content editables
This introduces a view helper ``contentElement.wrap`` which allows for
explicitly wrapping template parts with node meta data that is required
for the backend to show properties in the inspector.

Usually this ViewHelper is not required, but it enables usage of the
``contentElement.editable`` ViewHelper outside of content element
templates.

This is especially useful if you want to make properties of a custom
document node inline-editable::
 <neos:contentElement.wrap>
   <div>
     {neos:contentElement.editable(property: 'someProperty')}
   </div>
 </neos:contentElement.wrap>

Change-Id: I89db47195a384f01a8329f8a616a70d086a82c96
Reviewed-on: https://review.typo3.org/26049
Reviewed-by: Bastian Waidelich
Tested-by: Bastian Waidelich