[TASK] Improve backward compatibility of Site Import
when attempting to import "old" Site XML, there is the need
to make the site node a Shortcut to the first subpage, flag it
to be hidden in index, and add a `title` property being the
site's name.

This introduces a method to "patch" the incoming XML to
comply with the above requirements.

Note: this reverts parts of Iddc86edb51df20f1c72e280f8571b918a09af0f6
where the backward compatibility requirements were already
partially implemented.

Change-Id: I3280096189f5f5eafa7b9c684d3d64c9376fe0f1
Resolves: #53609
Related: #53381
Reviewed-on: https://review.typo3.org/26071
Reviewed-by: Adrian Föder
Tested-by: Adrian Föder
Tested-by: Mr. Hudson
Reviewed-by: Karsten Dambekalns
Reviewed-by: Bastian Waidelich
Tested-by: Bastian Waidelich