[TASK] Implement TypoScript AutoInclude setting for TypoScriptView
Packages can now register with the setting::

  TYPO3:
    Neos:
      typoScript:
        autoInclude:
          'MyVendor.MyPackageKey': TRUE

to get TypoScript in the path:
``MyVendor.MyPackageKey/Private/TypoScript/Root.ts2``
included automatically.
The order of inclusions is set by the package loading order
(and thus dependency chain of your package).
This also allows disabling of autoIncludes if needed.

Additionally it cleans the interface of methods in TypoScriptView
and TypoScriptService and prevents two parsing runs of the
TypoScript by caching the Runtime instance in the view.

Finally removes all references to the NodeTypes package.

Releases: 1.0, master
Change-Id: Iadb1bcf48d7fca1f04a343e15c03002e44bc6502
Reviewed-on: https://review.typo3.org/26161
Reviewed-by: Christian Mueller
Tested-by: Christian Mueller
Reviewed-by: Christopher Hlubek
Tested-by: Christopher Hlubek