[TASK] Integrate TypoScript content cache in Neos
This change implements a flushing strategy on node updates that will
flush entries in the content cache with matching entry tags.

For each node, a set of tags will be derived on node changes (add,
update, remove):

- "Everything" will always be included
- "Node_[Identifier]" for the node itself
- "NodeType_[NodeType]" for all node types that the node implements
- "DescendantOf_[Identifier]" for all parents until the root node

By using these tags in the "@cache.entryTags" configuration, a clever
flush strategy can be implemented depending on the content and structure
of a site.

A @cache configuration is added to the "root" path and the
TYPO3.Neos:Page prototype that should provide a reasonable default.
The TYPO3.Neos:Plugin prototype and the Neos backend parts of the page
are marked as uncached for proper caching behaviour.

Depends on Ib202a8222655685982b793b19e3bebf2eb7ba99f to store the node
type names as cache tag values.

Change-Id: Ia83bd0dc6ba50150463b84d24f546d1da682a048
Releases: master
Reviewed-on: https://review.typo3.org/27491
Tested-by: Robert Lemke
Reviewed-by: Robert Lemke