[BUGFIX] Shortcut to siteroot has no active state
The new menu implementation returns a state normal for
nodes equal to the siteroot. This prevents a menu which
contains a shortcut to the homepage to render an active
state.

A menu could now render an active state on the shortcut
to the homepage by for example using a custom node type
for the homepage and adding the following condition:

  <f:if condition="{item.node.nodeType.name} == 'My.Package:Homepage'">
      <f:then>
          <li class="{f:if(condition:'{item.state}==\'current\'', then:'active')}">
              <neos:link.node node="{item.node}">{item.label}</neos:link.node>
          </li>
      </f:then>
      <f:else>
          ...
      </f:else>
  </f:if>

Change-Id: Ia44e2ef7e8ecb3713f38fa2adb5256398792c64d
Related: #55873
Releases: master, 1.0
Reviewed-on: https://review.typo3.org/27541
Reviewed-by: Rens Admiraal
Tested-by: Rens Admiraal
Reviewed-by: Christian Mueller
Tested-by: Christian Mueller
Reviewed-by: Karsten Dambekalns
Tested-by: Karsten Dambekalns