[FEATURE] AssetEditor for inspector
This change adds an editor and updates Neos to support two new property
types: TYPO3\Media\Domain\Model\Asset for a single reference to an Asset
entity and array<TYPO3\Media\Domain\Model\Asset> to select or upload
multiple assets.

The ContentElementWrappingService was changed to export all public
properties only for the ImageVariant type and support for the Asset
types was added. This part of Neos needs more refactoring in the future.

Change-Id: I754f92cb9f8b458931938a8562551fe48a6518eb
Releases: master, 1.1
Reviewed-on: https://review.typo3.org/27597
Reviewed-by: Rens Admiraal
Tested-by: Rens Admiraal
Reviewed-by: Christopher Hlubek
Tested-by: Christopher Hlubek