[BUGFIX] Prevent exception when matching incoming context URIs
This fixes a regression that was introduced with a previous commit
which resulted in an exception when logging into the backend.

Background:
A previous commit (Iea1b92b07daba69f5f76cd7ceb7b81f2d14d68a1)
streamlined the mapping of request paths to a node.
Unfortunately this introduced a regression that resulted in an
exception being thrown whenever a context URI (containing the
"@workspace" part) was matched by the homepage route.

This change fixes this by testing whether the given context path
could be matched.

Change-Id: Icd3be53d5d8abd44701ccacfbdd471090b88450c
Related: #54187
Releases: master
Reviewed-on: https://review.typo3.org/27688
Tested-by: Bastian Waidelich
Reviewed-by: Marc Neuhaus
Tested-by: Marc Neuhaus
Reviewed-by: Bastian Waidelich