[BUGFIX] Sites management module widget action button
The action button for the sites management module widget
points to a unresolved path due to the action name is camelcased.
Since all module paths are automatically lowercased we can safely
convert actions names into lowercase when creating links for them.
In case a custom path is desired that's already configurable.

Change-Id: I7c0823354802665cbb272c4a8475da31bfd83a68
Releases: master, 1.0
Reviewed-on: https://review.typo3.org/25381
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann
Reviewed-by: Mattias Nilsson
Tested-by: Mattias Nilsson
Reviewed-by: Karsten Dambekalns
Tested-by: Karsten Dambekalns