[TASK] Only cache content collection in primary content per default
The content cache mode "cached" was enabled for all content collections
which results in a potentially large number of cache entries on pages
with nested elements (e.g. multi-column). This could reduce the effect
of the content cache.

With this change only the PrimaryContent prototype uses mode "cached"
while it has to be set explicitly on custom content collections.

Example::

	teaser = ContentCollection {
		nodePath = 'teaser'
		@cache {
			mode = 'cached'
		}
	}

Change-Id: I1291ea66cc61ef1f76ee90dc3ac68d0d0a72589d
Releases: master
Reviewed-on: https://review.typo3.org/28040
Reviewed-by: Robert Lemke
Tested-by: Robert Lemke
Reviewed-by: Christopher Hlubek
Tested-by: Christopher Hlubek