[FEATURE] Linking to assets in Aloha editor
This change introduces the possibility to link to files within
text elements. The autocomplete wizard of the link wizard now
displays pages and assets in the result list.

The ConvertNodeUrisImplementation is also renamed to
ConvertUrisImplementation to reflect that it does convert uris
and not only node uris.

Aloha's table wizard and link wizard are also moved below
the #neos-application element for better separation of styling.

Change-Id: I40db54541bc990a6620d2b0c4180f7a8646061c6
Resolves: NEOS-20
Releases: 1.1, master
Reviewed-on: https://review.typo3.org/28696
Reviewed-by: Rens Admiraal
Tested-by: Christopher Hlubek
Reviewed-by: Christopher Hlubek