[FEATURE] Localize the UserInterface
Implemented Ember-i18n in order to localize the Handlebar templates
and Javascript strings. Translations are read from the XLIFF files
and two new settings are added under TYPO3.Neos.userInterface:

- locale: the default locale for the backend interface
- availableLocales: the locales available in the system

For the Handlebars template, the helper works similar like the Fluid
Translate ViewHelper, so it parses the string based on the key, or
renders the key if the key is not found::

	{{translate id="key"}}

In case no translation is found for the key, the default value  is
returned as output. In case a translation is found for the key,
that is returned::

	{{translate id="key" value="default content"}}

Alternative syntax::

	{{#translate id="key"}}content{{/translate}}

Possible arguments to `translate`:

- `id`, the identifier (required)
- `value`, default content
- `package`, packageKey (defaults to TYPO3.Neos)
- `source`, sourceCatalogue (defaults to Main)

For the `id` dots (`.`) inside the xliff file will be ported to
dashes (`-`) in order to have programmatically correct json files
without losing functionality in the xliff files.

I.e::

	my.xliff.label.identifier

becomes::

	my-xliff-label-identifier

Adds an interface language selector to the user preferences in the
Administration/UserModule. The preference is saved to the users
preferences.

Adds methods to the UserService to fetch the user preferences.

Change-Id: I87e34e603027682112b53874c050b9dd16e52749
Releases: master
Related: NEOS-61, NEOS-89, NEOS-143, NEOS-435, NEOS-436, NEOS-446