[FEATURE] Module Users: Allow arbitrary roles to be assigned to accounts
Previously the user management module only allowed for assigning one
of the roles ``TYPO3.Neos:Editor`` or ``TYPO3.Neos:Administrator`` to
be assigned to a user.

With this change all configured roles that are not abstract can be
assigned to a user, also multiple ones.

Furthermore this change contains some (mostly cosmetic) tweaks in PHP
code and Fluid templates.

Change-Id: I0ca000c233e36a2e9c3689c1dd405033274955b3
Resolves: NEOS-115
Resolves: NEOS-543
Releases: master