[BUGFIX] Asset editor calls server without identifier
When the list of assets allows multiple assets and has
no value set it will send a json encoded empty array
to the asset metadata endpoint. This results in an internal
server error that is thrown to the user without further
reason as we can just remove the loading state.

Releases: master, 1.1
Change-Id: Id875c5d210495645743705bd843ebbf15db3d0ba
Reviewed-on: https://review.typo3.org/30530
Reviewed-by: Christian Mueller
Tested-by: Christian Mueller
Reviewed-by: Aske Ertmann