[BUGFIX] Publishing nodes gives a server error
Publishing changed and removed nodes gives a internal
server error because of an incorrect property mapping
configuration. Besides that possible validation errors
resulted in a 403 access denied as the errorAction on the
AbstractServiceController was not added to the Policy.yaml

Since I304e0b76164f83a85c3ca3f4d0ed630b04319426 the
convertNodes method is also obsolete and removed in this
change.

Change-Id: Idbda6ab00b60fad51b22ebc647e170364b99319d
Releases: master
Reviewed-on: https://review.typo3.org/30744
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann