[!!!][BUGFIX] Menu state should not be calculated based on a shortcut
Fixes wrong behavior of menu states, introduced with the refactored
MenuImplementation. Especially behavior of shortcuts pointing to a
subpage of themself showed wrong current states.

This is only breaking compared to the 1.1 branch not to 1.0.
If you need a link to the site root page with correct active state
you shouldn't create a shortcut pointing to it but rather include
the real site root in the menu, by using the itemCollection
property of the Menu TypoScript object.

TypoScript example::

  itemCollection = ${q(site).add(q(site).children('[instanceof TYPO3.Neos:Document]')).get()}

Change-Id: I4c0a92e3bf068b53c5d0286248feb6e8236894a7
Releases: master, 1.1
Reviewed-on: https://review.typo3.org/30848
Reviewed-by: Christian Mueller
Tested-by: Christian Mueller
Reviewed-by: Rens Admiraal
Tested-by: Rens Admiraal