[TASK] Refactor site:* services to use TYPO3CR services
This changes the implementation of the site:export and site:import
service to use the new TYPO3CR node export / import domain services
instead. Only the "site" specific part remains in Neos.

If the "legacy" (1.0 and 1.1) Sites.xml format is detected, the old parser
is used (which has been renamed to LegacySiteImportService).

Change-Id: Ice370b45bb924e9065d4a65e3ac976d1d0e76729
Resolves: NEOS-166
Releases: master