[BUGFIX] Behat test "Toggle Preview Mode" is broken
The "Toggle Preview Mode" test is broken as the "Full Screen" button
might not yet be rendered during the test. This change introduces
a new @When method to the Behat feature context to support waiting
for an element to be rendered.

Usage::

  When I wait for the "Open full screen" button to be visible

The element should be added to the switch statement in the
namedElementSelectorExists method of the FeatureContext.

Change-Id: I2fb1d3c9f2fd5dccdc148e1a54cb26ccd316aa7d
Fixes: NEOS-173
Releases: master, 1.1, 1.0