[FEATURE] Time support for the inspector date editor
Add support for date formats with time in the inspector
date editor. To support this the format is changed to include
time and timezone offset, to ensure consistency between
server and client.

Example configuration::

  editorOptions:
    format: 'dd-mm-yyyy hh:ii'

Additionally fixes a bug where the date isn't reset when
the inspector changes are discarded.

Change-Id: Ic9d371a498c0ff63b8115be0835625bf585bfe10
Resolves: NEOS-203
Releases: master