[!!!][FEATURE] Switch to PHP date format for inspector date editor
Switch from a custom date format used in the inspector date editor
to the standard PHP date format. This is a format that more people
are familiar with so it makes sense to use it instead. It also
makes the format indenpendent of the date edtior implementation.

Example configuration::

  editorOptions:
    format: 'd-m-Y H:i'

This is marked breaking due to the format change, meaning that if you
have changed the default date format you have to migrate to the new
one.

To adjust the code use the new format, it should be enough to run this
on your site package(s)::

      ./flow flow:core:migrate --package-key <sitepackagekey>

Change-Id: I3c6f960839bd751508a1c1c0e0c9aa327c5ed4cd
Resolves: NEOS-205
Releases: master