[BUGFIX] Backend redirects breaks with invisible/inaccessible nodes
If a node is inaccessible or invisible (hidden after/before) the backend
redirect service cannot find it thus leading to not being able to
redirect to the page.
In case the fallback redirect is used to the site's root node and the root
node is inaccessible/invisible the backend will throw a 404, even though
the page can be accessed just fine in the content module.

This is caused because the context created for finding the nodes in the
backend redirect service doesn't allow inaccessible/invisible nodes.

Change-Id: I6af906f19f3e351354b1d5568c96e688fbaaf4a2
Releases: master, 1.1, 1.0
Fixes: NEOS-225