[FEATURE] Apply constraints to new/paste buttons in navigate component
Applies constraints to the new/paste buttons in the navigate component
trees (node tree/context structure tree). This is done by checking if
the selected node allows children or siblings and if so allows the
matching position. In case of pasting the cut/copied node is checked
against the allowed children/siblings. If there's no available position
the buttons are disabled. If the chosen position is not available the last
available option is chosen by default (into/after). Takes the unmodifiable
levels configuration into account.

Change-Id: I32ee139a3bb8d63e1d8a1a7c13c795a85698d8cf
Releases: master
Resolves: NEOS-187
Resolves: NEOS-159
Related: NEOS-215