[TASK] Use placeholder instead of default value in inline link editor
When creating a new link the link field now holds a placeholder
instead of the default value "http://" explaining that the field
can be used for links and searching for nodes/assets. If a link is
inserted without http:// it is prefixed automatically if it matches
an external link pattern, to avoid broken links.

Change-Id: Ic3d9cea6a82c4f3c7850a1370f879f5ab5c4594a
Releases: master
Resolves: NEOS-232