[FEATURE] Rudimentary RESTful Nodes service
This change refactors and introduces more functionality for a RESTful
service providing access to nodes in Neos. The preferred output format
is HTML, but JSON is also supported (triggered via Accept header).

The existing NodeController was renamed to NodesController and references
to it were corrected accordingly.

Change-Id: Id7f57fe9fa2b94dbe4955ed47ffdf651be71e678
Related: NEOS-199
Releases: master