[FEATURE] Basic RESTful Assets service
This change refactors and introduces more functionality for a RESTful
service providing access to assets in Neos. The preferred output format
is HTML, but JSON is also supported (triggered via Accept header).

The existing AssetController was renamed to AssetsController and
references to it were corrected accordingly.

The change also fixes the JSON output for the Nodes service.

Change-Id: Idd6b2893c244375c79b3ffbb324698cbb33660cc
Resolves: NEOS-200
Related: NEOS-199
Releases: master