[TASK] Swap some access checks with workspace checks
This change inverts a number of checks for the workspace being not the
live workspace and for access to the TYPO3_Neos_Backend_GeneralAccess
resource.

This makes it possible to render full Neos pages in a CLI context (where
the security framework cannot be initialized) and should speed up things
a tiny bit along the way (by avoiding security checks in case of live
being rendered).

Change-Id: Ibe30042ec079ececf7441295dd21ffd9b0259760
Releases: master, 1.1