[TASK] Rename node:autocreatechildnodes command to node:repair
This change renames the node:autocreatechildnodes to node:repair and
by that creates a command which can perform further "repair" tasks in
the future. From a user's perspective simply running "repair" will likely
solve all structural node problems.

Further more, the command controller is moved to the TYPO3CR package
and thus consists of two commits: this one which removes the controller
from TYPO3.Neos and a second one which adds it to TYPO3CR.

Change-Id: Iab0c00fc18bb612e622860bc892611f26ac38b5d
Releases: master
Related: NEOS-325