[FEATURE] Improved workspace:* commands
This change renames and improves the workspace:publishall
and workspace:discardall commands. The new commands
workspace:publish and workspace:discard accept a new
argument "--dry-run" which allows for simulating publishing
and discarding of the given workspace.

The old commands are still available for backwards
compatibility. They are now marked as deprecated.

Change-Id: I26f49353ac470b04a08aee4721c5c8de8bd371d0
Releases: master