[!!!][FEATURE] Make button label and highlighting mode configurable
The HTML editor uses code mirror and could thus be used to edit
other types of textual content as well. To allow this two new
``editorOptions`` are introduced:

- ``buttonLabel`` overwrites the text on the button to open the editor
- ``highlightingMode`` overwrites the default (text/html) highlighting

This is breaking if you used the HtmlEditor in your node types. To
adjust, you need to replace *HtmlEditor* with *CodeEditor*, either
manually or with::

  ./flow core:migrate --package-key <your package>

Change-Id: I134d0cf4de2ed155faa0088f3dfa2eb406f662c0
Releases: master, 1.2
Resolves: NEOS-379