[BUGFIX] Content collection falls back to first child if nodePath is empty
When a content collection cannot find a node in the given path, it falls
back to the first found child node on the current page. This leads to very
unexpected and confusing results, e.g. rendering the main content collection
right before the content collection with the unfound node.

Change-Id: Ib9be0478a1003ff708b00cc55a2608d76c1ddb9c
Releases: master, 1.1, 1.0