[!!!][TASK] Deprecate unused IncludeJavaScriptViewHelper
This VH has been mostly untouched since it's inception and is no longer
used. The best practice for including JS is now a different one: either have
your scripts in a section of your template or add them to the TS Array in
page.head.javascripts in your TypoScript.

Thus this VH is deprecated and will be removed in three versions from now.

Change-Id: Ifd0f0aa279e860e95cc37dda0519d42695a3a792
Releases: master