[BUGFIX] Respect inactive state for domains
If a domain is inactive the domain will not be used for determining the
active site. This means that when requesting an inactive domain the
default site will be shown instead, as if the domain hasn't been
configured.

Previously the domain state wasn't taken into account at all meaning
that inactive domains would still match their respective site.

Change-Id: Icc69889c0713faf7e5e4a51429cbc1ea64497f27
Releases: master, 1.2
Fixes: NEOS-181