[!!!][TASK] Move FlowQueryOperations concerning nodes to TYPO3CR
All FlowQuery operations concerning nodes have no dependencies
on the TYPO3.Neos package, but only on the TYPO3.TYPO3CR package.
The rightfully belong there and probably ended up in the Neos
package before the TYPO3.Eel package became part of the Flow base
distribution. Now that Eel is available all the time the
operations can as well reside in TYPO3CR which allows to use
them without installing the Neos package at all.

This change is breaking if you extended the FlowQueryOperations
that are now moved. You need to extend from the same class in
TYPO3CR now.

The needed namespace replacement is from::

  ``TYPO3\Neos\TypoScript\FlowQueryOperations``

to::

  ``TYPO3\TYPO3CR\Eel\FlowQueryOperations``

Change-Id: Ifa2753adaacf6244b97ead192bd8ad45a5b7e37f
Releases: master
Resolves: NEOS-405