[BUGFIX] Publishing the page after moving a content element broken
After moving a content element on a page and publishing using the
"Publish" button, a "Conflict" error is thrown. This is due to the
shadow node being rendered and being included as a unpublished node,
but when trying to publish it cannot find it since it's removed.

Change-Id: I27728de88cf3bfbdfe24de131737a3cd90153f61
Related: NEOS-389
Releases: master