[TASK] Improve 404 handling in Neos backend
Currently if the editor discards a page that doesn't exist in the live
workspace, Neos returns a server error.

This patch solves the problem by returning a proper 404 error and handle
this error in the HttpClient. Neos will try to go up in the tree until
it find a valid document to display.

This patch also introduces a new aspect to add the current node
identifier as a cache tag. This allows more flexible cache invalidation
in Neos.

Change-Id: I154ea045e9e4a23a96406b1193dc99ec35c28f4f
Resolves: NEOS-138
Releases: master