[FEATURE] Routing: content dimension / translation support
This change introduces translatable URI segments and
general support for content dimensions in the frontend node
route part handler. Editors may change the path segment of
a specific document node (for example a page) in the inspector.

The route part handler which renders and matches the URIs
does not use a node's name anymore but refers to a new
node property "uriPathSegment". This finally decouples
URI generation from the internal node structure and naming.

This change also makes sure that router caches are flushed if
any node is updated, removed, published or discarded. That
is refined in a separate change.

In order to migrate existing websites, you must run the "node:repair"
command in order to fill the new uriPathSegment properties.

If you used the "LocalizedRoutePartHandler" from a development
version of Neos, you need to remove the respective Objects.yaml
configuration.

Change-Id: I80228052f82aaf6c649530096e07f364cd9191ba
Releases: 1.2, master
Resolves: NEOS-296