[FEATURE] Crop aspect ratio lock presets (image editor)
Introduces aspect ratio lock presets for the cropping
in the inspector image editor. Multiple presets can be
configured, additionally it's optional to allow custom
ratios and to enable the original ratio as an option.

Additionally a locked aspect ratio can be configured
to ensure a specific aspect ratio of all images in
that image property.

Example configuration::

  editorOptions:
    crop:
      aspectRatio:
        options:
          square:
            width: 1
            height: 1
            label: 'Square'
          fourFive:
            width: 4
            height: 5
        enableOriginal: TRUE
        allowCustom: TRUE
        locked:
          width: 0
          height: 0

Example of disabling available options::

  # Disable ``square`` option
  'Acme.Demo:Image':
    properties:
      image:
        ui:
          inspector:
            editorOptions:
              crop:
                aspectRatio:
                  options:
                    square: ~

Change-Id: Iabf105bb4fd2add0e800a52124f01f27b1fe117e
Releases: master, 1.2
Resolves: NEOS-557