[BUGFIX] Deselecting image elements leads to an error
When deselecting an element containing an image property,
the request for the image is aborted if it's still in progress.

However the abort method was removed in
I10fab119aeab6f6e2e86a217ba71c6f2b5c1efe0 as part of a cleanup.

This change reintroduces the abort method, preventing the error
from happening.

Change-Id: I8f8ed795390d7e8d02b20bc9937147c14749f357
Fixes: NEOS-555
Releases: master