[BUGFIX] Gracefully handle links to removed nodes/assets
Links to nodes/assets that cannot be resolved will now be
unlinked automatically.

When nodes and assets can not be resolved in the LinkService an
exception is thrown. This causes an issue in the ConvertUrisImplementation
object as it will throw an exception. This leads to the fatal
error ``__toString should not throw an exception`` and as such
a white page on production servers.

This change makes LinkingService::createNodeUri() and createAssetUri()
return NULL if $node/$asset was NULL as well (which was already
anticipated by the method anyways).

Change-Id: I208adbc6d812f3ef357903844eb044a1065e75eb
Fixes: NEOS-564
Releases: master, 1.2