[FEATURE] Display modal dialog for initial translation of a document
If a document does not exist in the target language, the editor can
choose whether content from the current language should be copied to
the target language, or whether he'd like to start from scratch.

Additionally, all documents which do not exist in the rootline in
the target variant are also materialized.

Change-Id: I05a386819f2f62790b6339e768b5e66193371d5b
Resolves: NEOS-633
Resolves: NEOS-646
Releases: master