[BUGFIX] Routing: only flush related caches when a node changes
Currently the complete routing cache is flushed whenever a node is
changed. This can have a noticeable negative impact on performance of
both Frontend and Backend.

With this change only the cache for affected nodes is cleared when
a node is published and/or the "uriPathSegment" property was changed.

Change-Id: Ic131d828f491b8559d93647625298a5eccaeb131
Depends: I49da56cd705e41fdc96fe84ac4936d268a0b6f06
Releases: 1.2, master
Fixes: NEOS-673
Related: NEOS-296