[BUGFIX] Exception instead of 404 in certain cases
Neos responds with a "page not found" screen in some cases. If a page
cannot be found due to "normal" reasons, this works fine. In some cases,
the reason for the page not found is not a node that does not exist.

So when in a multi language site (like the demo site) the language
segment is removed, the rendering of the error page takes a different
path and the Backend\TranslateViewHelper is involved.

That uses the security context without checking if it can be initialized,
leading to an exception.

This change adds a check using canBeInitialized() to that.

Change-Id: I88ddd922a9c515aac1e3a4771695ef54870ec02a
Releases: master