[TASK] Use "language" instead of "locale" and enable UI languages
The current implementation of UI localization support uses
the term "locale" when referring to the user interface
language. That's not completely wrong, since we use a locale
to identify the language, but from a user's POV what we really
mean is the language.

Once we support date / time formats etc, we'll most likely
introduce more settings so the user can choose this
independently from the language.

Additionally we currently only provide "English" in the list
of UI languages. Let's add some more we are sure of being
available until the preview release.

Change-Id: I78b07a01f3c9f2304db35b2ee7990c6a58b7517c
Resolves: NEOS-769
Release: master