[TASK] Don't log "node discarded" events
This changes removes logging of discard operations which undo changes
in an editor's local workspace. Discard operations are not generally
of interest to other editors so they can be removed from the global
history for now. Additionally, if a discard event has been logged in
the past, the History module would fail on rendering the list of events
since the Node Discard case hasn't been implemented yet.

Change-Id: Ia60793186283233e1fd5b2e41317588c235138ba
Resolves: NEOS-914
Releases: master