[BUGFIX] "uriPathSegment" not created via auto-created child nodes
Fixes the issue where the property "uriPathSegment" is missing when
document nodes are created via the auto-create child nodes mechanism.

If a node type triggers creation of a sub node of type
"TYPO3.Neos.NodeTypes:Page", that new node does not contain a
"uriPathSegment" property and thus results in exception, for example if
the sub document nodes are shown in the node tree.

The solution provided by this change registers a simple slot for the
Node Added signal which sets the uriPathSegment property to the
node name if that property does not exist yet.

Change-Id: I1e399a6dcb713989e01a5890dce61751cf9a9459
Resolves: NEOS-925
Releases: 1.2, master