[BUGFIX] Linking via context path works in non TypoScript environment
Since 1.2 you couldn't link to a node using the context node path outside
of TypoScript environments. This feature can be useful in lots of
use-cases where you have a node path but no TypoScript context.

This change fixes that behavior and adds tests that actually test this
to prevent this kind of regression.

Change-Id: Ib7acb494f96459043eba9d33ef129e48ed9c0fb1
Releases: master, 1.2
Fixes: NEOS-961